Move navigation dropdown components
