+++ src/components/Navigation/SiteHeader.vue
@@ -54,13 +54,13 @@
 <script>
 import { mapState } from 'vuex';
 import i18n from '@/locale/i18n';
-import SiteSidebar from '@/components/Navigation/SiteSidebar.vue';
-import UserProfileMenu from '@/components/UserProfileMenu.vue';
+import SiteSidebar from '@/components/Navigation/SiteMenuDropdown.vue';
 import getActiveItemType from '@/util/getActiveItemType';
 import NavMenuItem from '@/components/Navigation/navbar/NavMenuItem.vue';
 import NavMenu from '@/components/Navigation/navbar/NavMenu.vue';
 import PeriodSelector from '@/components/Navigation/toolbar/PeriodSelector.vue';
 import ViewToggle from '@/components/Navigation/toolbar/ViewToggle.vue';
+import UserProfileMenu from '@/components/Navigation/UserMenuDropdown.vue';
 
 export default {
   name: 'SiteHeader',
